+++ src/main/java/com/fitech/app/users/domain/services/impl/UnitOfMeasureServiceImpl.java
@@ -3,11 +3,10 @@ package com.fitech.app.users.domain.services.impl;
 import com.fitech.app.commons.util.PaginationUtil;
 import com.fitech.app.users.application.dto.UnitOfMeasureDto;
 import com.fitech.app.users.application.exception.UnitOfMeasurerNotFoundException;
-import com.fitech.app.users.application.exception.UserNotFoundException;
 import com.fitech.app.users.application.wrappers.ResultPage;
 import com.fitech.app.users.domain.entities.UnitOfMeasure;
 import com.fitech.app.users.domain.services.UnitOfMeasureService;
-import com.fitech.app.users.infrastructure.repositories.UnitOfMeasureRepository;
+import com.fitech.app.users.infrastructure.repository.UnitOfMeasureRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -46,7 +45,7 @@ public class UnitOfMeasureServiceImpl implements UnitOfMeasureService {
 
     @Override
     @Transactional
-    public UnitOfMeasureDto update(Long id, UnitOfMeasureDto unitOfMeasureDto) {
+    public UnitOfMeasureDto update(Integer id, UnitOfMeasureDto unitOfMeasureDto) {
         UnitOfMeasure existingUnit = unitOfMeasureRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Unidad de medida no encontrada"));
 
@@ -69,7 +68,7 @@ public class UnitOfMeasureServiceImpl implements UnitOfMeasureService {
 
     @Override
     @Transactional
-    public void delete(Long id) {
+    public void delete(Integer id) {
         if (!unitOfMeasureRepository.existsById(id)) {
             throw new EntityNotFoundException("Unidad de medida no encontrada");
         }
@@ -78,7 +77,7 @@ public class UnitOfMeasureServiceImpl implements UnitOfMeasureService {
 
     @Override
     @Transactional(readOnly = true)
-    public UnitOfMeasureDto findById(Long id) {
+    public UnitOfMeasureDto findById(Integer id) {
         UnitOfMeasure unitOfMeasure = unitOfMeasureRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Unidad de medida no encontrada"));
         return convertToDto(unitOfMeasure);
